% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{compute_relative_expression}
\alias{compute_relative_expression}
\title{Compute and Visualize Relative Gene Expression Per Cell}
\usage{
compute_relative_expression(seurat_obj, output_path, top_gene = 20)
}
\arguments{
\item{seurat_obj}{A Seurat object containing single-cell RNA-seq data.}

\item{output_path}{Character; the directory path where the output PNG image will be saved.}

\item{top_gene}{Integer; the number of top expressed genes to visualize. Default is 20.}
}
\value{
None (the function generates and saves a boxplot). The function also prints the output file path
  and the gene with the highest variation among the selected top genes.
}
\description{
This function computes the relative expression of each gene per cell using sparse matrix operations
and generates a boxplot of the most expressed genes.
}
\examples{
# Example usage:
compute_relative_expression(seurat_obj, output_path = "../../results/01-QC", top_gene = 20)

}
