% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{read_gtf_file}
\alias{read_gtf_file}
\title{Read and Process a GTF File}
\usage{
read_gtf_file(
  gtf_file,
  gene_info_file = "zUMIs_annot.csv",
  quant_list,
  feature_type
)
}
\arguments{
\item{gtf_file}{Path to the GTF file.}

\item{gene_info_file}{Path to the gene info file (if available, it will be used instead of parsing the GTF file).}

\item{quant_list}{A list containing gene expression matrices for different samples.}

\item{feature_type}{Character vector specifying which feature types to extract from `quant_list`.}
}
\value{
A data frame containing processed gene annotations.
}
\description{
This function reads a GTF file, extracts gene annotations, and saves the processed data to a CSV file.
If a preprocessed gene info file exists, it is loaded instead. `data_genes` is derived dynamically
from `quant_list` based on the provided `feature_type`.
}
\examples{
gtf_data <- read_gtf_file("path/to/gtf.gtf", "path/to/gene_info.csv", quant_list, feature_type = c("exon", "intron"))
}
