% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{read_zUMIs_data}
\alias{read_zUMIs_data}
\title{Read zUMIs Data}
\usage{
read_zUMIs_data(
  metadata_file_name,
  metadata_dir,
  sample_name_column,
  data_path_column
)
}
\arguments{
\item{metadata_file_name}{Character. Name of the metadata file.}

\item{metadata_dir}{Character. Directory where the metadata file is located.}

\item{sample_name_column}{Character. Column name in the metadata file containing sample names.}

\item{data_path_column}{Character. Column name in the metadata file containing data paths.}
}
\value{
A list containing:
\item{expression_list}{List. Loaded expression data.}
\item{umicount_list}{List. Loaded UMI count data.}
\item{genecount_list}{List. Loaded gene count data.}
\item{wellbarcode_list}{List. Loaded well barcode data.}
\item{quant_list}{List. Exported expression matrix.}
}
\description{
This function reads expression, UMI count, gene count, and well barcode data from zUMIs output files
based on metadata information and exports the expression matrix.
}
\examples{
\dontrun{
data_list <- read_zUMIs_data(
  metadata_file_name = "metadata.csv",
  metadata_dir = "path/to/metadata",
  sample_name_column = "SampleID",
  data_path_column = "DataPath"
)
}

}
