% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{read_zUMIs_expression}
\alias{read_zUMIs_expression}
\title{Read zUMIs Expression Data}
\usage{
read_zUMIs_expression(metadata_path, sample_name_column, data_path_column)
}
\arguments{
\item{metadata_path}{The directory path where the metadata file is located.}

\item{sample_name_column}{The column name in the metadata file that contains sample identifiers.}

\item{data_path_column}{The column name in the metadata file that contains the path to sample data.}
}
\value{
A list containing the expression data for each sample.
}
\description{
Reads and loads zUMIs expression data from RDS files, based on metadata provided in a tab-separated file.
}
\examples{
expression_list <- read_zUMIs_expression("metadata.tsv", "path/to/metadata", "sample_id", "data_path")
}
