% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{read_zUMIs_genecounts}
\alias{read_zUMIs_genecounts}
\title{Read zUMIs Gene Counts}
\usage{
read_zUMIs_genecounts(metadata_path, sample_name_column, data_path_column)
}
\arguments{
\item{metadata_path}{Path to the metadata file (TSV format).}

\item{sample_name_column}{The column name in the metadata file that contains sample identifiers.}

\item{data_path_column}{The column name in the metadata file that contains the path to sample data.}
}
\value{
A named list where each element is a data frame of gene counts per sample.
}
\description{
Reads gene count files for each sample specified in the metadata and returns them as a list.
}
\examples{
gene_counts_list <- read_zUMIs_genecounts("metadata.tsv", "sample_id", "data_path")
}
