% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{read_zUMIs_UMIcounts}
\alias{read_zUMIs_UMIcounts}
\title{Read zUMIs UMI Counts}
\usage{
read_zUMIs_UMIcounts(metadata_path, sample_name_column, data_path_column)
}
\arguments{
\item{metadata_path}{Path to the metadata file (TSV format).}

\item{sample_name_column}{The column name in the metadata file that contains sample identifiers.}

\item{data_path_column}{The column name in the metadata file that contains the path to sample data.}
}
\value{
A named list where each element is a data frame of UMI counts per sample.
}
\description{
Reads UMI count files for each sample specified in the metadata and returns them as a list.
}
\examples{
umi_counts_list <- read_zUMIs_UMIcounts("metadata.tsv", "sample_id", "data_path")
}
