% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{read_zUMIs_wellbarcodes}
\alias{read_zUMIs_wellbarcodes}
\title{Read zUMIs Well Barcodes}
\usage{
read_zUMIs_wellbarcodes(metadata_path, sample_name_column, data_path_column)
}
\arguments{
\item{metadata_path}{Path to the metadata file (TSV format).}

\item{sample_name_column}{The column name in the metadata file that contains sample identifiers.}

\item{data_path_column}{The column name in the metadata file that contains the path to sample data.}
}
\value{
A named list where each element is a vector of well barcodes per sample.
}
\description{
Reads well barcode files for each sample specified in the metadata and returns them as a list.
}
\examples{
well_barcodes_list <- read_zUMIs_wellbarcodes("metadata.tsv", "sample_id", "data_path")
}
